requirements.txt
pycrypto
mitmproxy
nose
